+++ docs/docusaurus.config.ts
@@ -4,7 +4,7 @@ import type * as Preset from '@docusaurus/preset-classic';
 
 const config: Config = {
     title: 'MapLibre Flutter',
-    tagline: 'Unofficial MapLibre bindings for Flutter',
+    tagline: 'Permissive and performant mapping library that supports Mapbox Vector Tiles (MVT) powered by MapLibre SDKs.',
     favicon: 'img/favicon-32x32.png',
 
     // Set the production url of your site here
@@ -71,7 +71,7 @@ const config: Config = {
                 },
                 {
                     type: 'docsVersionDropdown',
-                    position: 'right',
+                    position: 'left',
                 },
                 {
                     href: "https://pub.dev/documentation/maplibre/latest/maplibre/maplibre-library.html",