+++ example/lib/menu_page.dart
@@ -17,11 +17,15 @@ import 'package:maplibre_example/layers_hillshade_page.dart';
 import 'package:maplibre_example/layers_line_page.dart';
 import 'package:maplibre_example/layers_raster_page.dart';
 import 'package:maplibre_example/layers_symbol_page.dart';
+import 'package:maplibre_example/offline_page.dart';
 import 'package:maplibre_example/parameters_page.dart';
+import 'package:maplibre_example/permissions_page.dart';
 import 'package:maplibre_example/styled_map_page.dart';
 import 'package:maplibre_example/two_maps_page.dart';
+import 'package:maplibre_example/user_interface_page.dart';
 import 'package:maplibre_example/user_location_page.dart';
 import 'package:maplibre_example/web_controls_page.dart';
+import 'package:maplibre_example/widget_layer_page.dart';
 
 class MenuPage extends StatelessWidget {
   const MenuPage({super.key});
@@ -36,7 +40,7 @@ class MenuPage extends StatelessWidget {
         slivers: [
           const SliverToBoxAdapter(child: SectionTitle('General')),
           SliverGrid.extent(
-            maxCrossAxisExtent: 200,
+            maxCrossAxisExtent: 150,
             childAspectRatio: 1.5,
             children: const [
               ItemCard(
@@ -86,17 +90,39 @@ class MenuPage extends StatelessWidget {
                   iconData: Icons.gps_fixed,
                   location: UserLocationPage.location,
                 ),
+              ItemCard(
+                label: 'User interface',
+                iconData: Icons.control_camera,
+                location: UserInterfacePage.location,
+              ),
+              if (!kIsWeb)
+                ItemCard(
+                  label: 'Offline',
+                  iconData: Icons.wifi_off,
+                  location: OfflinePage.location,
+                ),
+              if (!kIsWeb)
+                ItemCard(
+                  label: 'Permissions',
+                  iconData: Icons.key,
+                  location: PermissionsPage.location,
+                ),
             ],
           ),
           const SliverToBoxAdapter(child: SectionTitle('Annotations')),
           SliverGrid.extent(
-            maxCrossAxisExtent: 200,
+            maxCrossAxisExtent: 150,
             childAspectRatio: 1.5,
             children: const [
               ItemCard(
                 label: 'Mixed Annotations',
                 iconData: Icons.control_point_duplicate_outlined,
                 location: AnnotationsMixedPage.location,
+              ),
+              ItemCard(
+                label: 'Widgets',
+                iconData: Icons.location_on,
+                location: WidgetLayerPage.location,
               ),
               ItemCard(
                 label: 'Circles',
@@ -122,7 +148,7 @@ class MenuPage extends StatelessWidget {
           ),
           const SliverToBoxAdapter(child: SectionTitle('Layers')),
           SliverGrid.extent(
-            maxCrossAxisExtent: 200,
+            maxCrossAxisExtent: 150,
             childAspectRatio: 1.5,
             children: const [
               ItemCard(
@@ -193,7 +219,7 @@ class ItemCard extends StatelessWidget {
         onTap: () => Navigator.of(context).pushNamed(location),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
-          children: [Icon(iconData), Text(label)],
+          children: [Icon(iconData), Text(label, textAlign: TextAlign.center)],
         ),
       ),
     );
@@ -207,6 +233,8 @@ class SectionTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListTile(title: Text(label, style: const TextStyle(fontSize: 18)));
+    return ListTile(
+      title: Text(label, style: const TextStyle(fontSize: 18)),
+    );
   }
 }