+++ lib/src/map_controller.dart
@@ -1,16 +1,38 @@
 import 'dart:typed_data';
 
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
 import 'package:maplibre/maplibre.dart';
+import 'package:maplibre/src/inherited_model.dart';
 
 /// The [MapController] can be used to control, update and manipulate a
 /// rendered [MapLibreMap].
 abstract interface class MapController {
+  /// Find the [MapController] of the closest [MapLibreMap] in the widget tree.
+  /// Returns null if called outside of the [MapLibreMap.children].
+  static MapController? maybeOf(BuildContext context) =>
+      MapLibreInheritedModel.maybeMapControllerOf(context);
+
+  /// Find the [MapController] of the closest [MapLibreMap] in the widget tree.
+  /// Returns null if called outside of the [MapLibreMap.children].
+  static MapController of(BuildContext context) =>
+      maybeOf(context) ??
+      (throw StateError('Unable to find an instance of MapController'));
+
   /// Convert a latitude/longitude coordinate to a screen location.
+  // TODO: can be made sync when flutter raster and ui thread are merged
   Future<Offset> toScreenLocation(Position lngLat);
 
   /// Get the latitude/longitude coordinate for a screen location.
+  // TODO: can be made sync when flutter raster and ui thread are merged
   Future<Position> toLngLat(Offset screenLocation);
+
+  /// Convert a latitude/longitude coordinate to a screen location.
+  // TODO: can be made sync when flutter raster and ui thread are merged
+  Future<List<Offset>> toScreenLocations(List<Position> lngLats);
+
+  /// Get the latitude/longitude coordinate for a screen location.
+  // TODO: can be made sync when flutter raster and ui thread are merged
+  Future<List<Position>> toLngLats(List<Offset> screenLocations);
 
   /// Instantly move the map camera to a new location.
   Future<void> moveCamera({
@@ -69,15 +91,23 @@ abstract interface class MapController {
   /// Get the current camera position on the map.
   MapCamera getCamera();
 
+  /// Get the current camera position on the map.
+  MapCamera? get camera;
+
   /// Returns the distance spanned by one pixel at the specified latitude and
   /// current zoom level.
   ///
   /// The distance between pixels decreases as the latitude approaches the
   /// poles. This relationship parallels the relationship between longitudinal
   /// coordinates at different latitudes.
+  // TODO: can be made sync when flutter raster and ui thread are merged
   Future<double> getMetersPerPixelAtLatitude(double latitude);
 
+  /// Get a list of all attributions from the map style.
+  Future<List<String>> getAttributions();
+
   /// The smallest bounding box that includes the visible region.
+  // TODO: can be made sync when flutter raster and ui thread are merged
   Future<LngLatBounds> getVisibleRegion();
 
   /// Add an image to the map.
@@ -85,6 +115,9 @@ abstract interface class MapController {
 
   /// Removes an image from the map
   Future<void> removeImage(String id);
+
+  /// Queries the map for rendered features.
+  Future<List<QueriedLayer>> queryLayers(Offset screenLocation);
 
   /// Show the user location on the map
   Future<void> enableLocation({