+++ lib/src/web/interop/interop.dart
@@ -11,6 +11,8 @@ part 'controls.dart';
 part 'events.dart';
 part 'gesture_handlers.dart';
 part 'map.dart';
+part 'map_geojson_feature.dart';
+part 'style_specification.dart';
 part 'marker.dart';
 
 /// A simple x/y [Point] class for JavaScript.