+++ lib/src/web/platform_impl.dart
@@ -1,6 +1,7 @@
-import 'package:flutter/widgets.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
-import 'package:maplibre/maplibre.dart';
+import 'package:maplibre/src/map_state.dart';
+import 'package:maplibre/src/offline/offline_manager.dart';
+import 'package:maplibre/src/permission_manager.dart';
 import 'package:maplibre/src/platform_interface.dart';
 import 'package:maplibre/src/web/widget_state.dart';
 
@@ -11,5 +12,14 @@ final class PlatformImpl extends PlatformInterface {
       PlatformInterface.instance = PlatformImpl();
 
   @override
-  State<MapLibreMap> createWidgetState() => MapLibreMapStateWeb();
+  MapLibreMapState createWidgetState() => MapLibreMapStateWeb();
+
+  @override
+  Future<OfflineManager> createOfflineManager() =>
+      throw Exception('The OfflineManager can not be used on web.');
+
+  @override
+  PermissionManager createPermissionManager() {
+    throw Exception('The PermissionManager can not be used on web.');
+  }
 }