+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.1.1
+
+### New Features
+
+- Add `controller.queryLayers()` to check what layers have been clicked.
+- Add scalebar, zoom buttons, compass and attribution widgets as a unified user
+  interface.
+- Improve hash distribution
+- Add `MapOptions` parameters to toggle platform specific UIs.
+
+### Bug Fixes
+
+- Fix `controller.getVisibleRegion()` throws exception on Android.
+
+Full
+Changelog: [v0.1.0...v0.1.1](https://github.com/josxha/flutter-maplibre/compare/v0.1.0...v0.1.1)
+
 ## 0.1.0
 
 ### Breaking Changes